Removed the viewpager, can now code and validate properly without worrying about anything. 09/06
