+++ app/src/main/java/com/example/lendahand/RegActivity1.java
@@ -1,0 +1,13 @@
+package com.example.lendahand;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class RegActivity1 extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_reg1);
+    }
+}
