+++ app/src/main/java/com/example/lendahand/RegActivity1Donee.java
@@ -1,0 +1,79 @@
+package com.example.lendahand;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.MotionEvent;
+import android.view.View;
+import android.widget.LinearLayout;
+
+import com.google.android.material.tabs.TabLayout;
+
+public class RegActivity1 extends AppCompatActivity {
+    private TabLayout tbDoneeReg;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        //hide title bar
+        getSupportActionBar().hide();
+        setContentView(R.layout.activity_reg1);
+        tbDoneeReg = findViewById(R.id.tbDonReg1);
+
+        LinearLayout tabStrip2 = ((LinearLayout) tbDoneeReg.getChildAt(0));
+        tabStrip2.getChildAt(1).setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return false;
+            }
+        });
+
+        LinearLayout tabStrip3 = ((LinearLayout) tbDoneeReg.getChildAt(0));
+        tabStrip3.getChildAt(2).setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
+        LinearLayout tabStrip4 = ((LinearLayout) tbDoneeReg.getChildAt(0));
+        tabStrip4.getChildAt(3).setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
+        LinearLayout tabStrip5 = ((LinearLayout) tbDoneeReg.getChildAt(0));
+        tabStrip5.getChildAt(4).setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
+        LinearLayout tabStrip6 = ((LinearLayout) tbDoneeReg.getChildAt(0));
+        tabStrip6.getChildAt(5).setOnTouchListener(new View.OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
+
+        tbDoneeReg.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
+            @Override
+            public void onTabSelected(TabLayout.Tab tab) {
+                switch (tab.getPosition()) {
+                    case 0:
+
+                }
+            }
+
+            @Override
+            public void onTabUnselected(TabLayout.Tab tab) {
+
+            }
+
+            @Override
+            public void onTabReselected(TabLayout.Tab tab) {
+
+            }
+        });
+    }
+}
