+++ app/src/main/java/com/example/lendahand/RegActivity4Donee.java
@@ -1,0 +1,16 @@
+package com.example.lendahand;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class RegActivity4 extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        //hide title bar
+        getSupportActionBar().hide();
+        setContentView(R.layout.activity_reg4);
+    }
+}
