+++ app/src/main/java/com/example/lendahand/RegActivityFinalDonee.java
@@ -1,0 +1,25 @@
+package com.example.lendahand;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+
+public class RegActivityFinal extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        //hide title bar
+        getSupportActionBar().hide();
+        setContentView(R.layout.activity_reg_final);
+    }
+
+    public void DoneeLogin(View view) {
+        //go to log in activity
+        Intent intent = new Intent(this, LoginScreenActivity.class);
+        startActivity(intent);
+        finish();
+    }
+}
